Convert `int` to `long` for pawn encoding indices across tablebase structures; remove redundant test assertions in `SyzygyImpProbeWDLIntegrationTest`.

+++ src/main/java/net/chesstango/piazzolla/syzygy/PawnDtz.java
@@ -116,7 +116,7 @@ class PawnDtz extends TableBase {
             i = ei.fill_squares(pos, flip, flip ? 0x38 : 0, p, i);
         }
 
-        int idx = ei.encode_pawn(p);
+        long idx = ei.encode_pawn(p);
 
         byte[] w = ei.precomp.decompress_pairs(idx);
