+++ src/main/java/net/chesstango/piazzolla/syzygy/PawnSymmetricWdl.java
@@ -79,7 +79,7 @@ class PawnSymmetricWdl extends TableBase {
             i = ei.fill_squares(pos, flip, flip ? 0x38 : 0, p, i);
         }
 
-        int idx = ei.encode_pawn(p);
+        long idx = ei.encode_pawn(p);
 
         byte[] w = ei.precomp.decompress_pairs(idx);
